+++ libp2p/discover.go
@@ -1,0 +1,58 @@
+package libp2p
+
+import (
+	"context"
+	"log"
+	"time"
+
+	dht "github.com/libp2p/go-libp2p-kad-dht"
+	"github.com/libp2p/go-libp2p/core/host"
+	"github.com/libp2p/go-libp2p/core/network"
+	"github.com/libp2p/go-libp2p/core/peer"
+	drouting "github.com/libp2p/go-libp2p/p2p/discovery/routing"
+	dutil "github.com/libp2p/go-libp2p/p2p/discovery/util"
+)
+
+func Discover(ctx context.Context, h host.Host, idht *dht.IpfsDHT, rendezvous string) {
+
+	var routingDiscovery = drouting.NewRoutingDiscovery(idht)
+	dutil.Advertise(ctx, routingDiscovery, rendezvous)
+
+	ticker := time.NewTicker(time.Second * 1)
+	defer ticker.Stop()
+
+	for {
+		select {
+		case <-ctx.Done():
+			return
+		case <-ticker.C:
+
+			peers, err := dutil.FindPeers(ctx, routingDiscovery, rendezvous)
+			if err != nil {
+				log.Fatal(err)
+			}
+			for _, p := range peers {
+				if p.ID == h.ID() {
+					continue
+				}
+				if h.Network().Connectedness(p.ID) != network.Connected {
+					_, err = h.Network().DialPeer(ctx, p.ID)
+					if err != nil {
+						continue
+					}
+				}
+			}
+		}
+	}
+}
+
+func getPeers(ctx context.Context, h host.Host, idht *dht.IpfsDHT, rendezvous string) ([]peer.AddrInfo, error) {
+
+	routingDiscovery := drouting.NewRoutingDiscovery(idht)
+	dutil.Advertise(ctx, routingDiscovery, "nunet")
+	peers, err := dutil.FindPeers(ctx, routingDiscovery, "nunet")
+	if err != nil {
+		panic(err)
+	}
+	return peers, nil
+}