+++ main.go
@@ -10,7 +10,7 @@ import (
 	_ "gitlab.com/nunet/device-management-service/docs"
 	"gitlab.com/nunet/device-management-service/firecracker"
 	"gitlab.com/nunet/device-management-service/internal"
-	"gitlab.com/nunet/device-management-service/internal/messaging"
+	"gitlab.com/nunet/device-management-service/libp2p"
 	"gitlab.com/nunet/device-management-service/routes"
 	"go.opentelemetry.io/otel"
 
@@ -19,7 +19,7 @@ import (
 )
 
 // @title           Device Management Service
-// @version         0.4.26
+// @version         0.4.27
 // @description     A dashboard application for computing providers.
 // @termsOfService  https://nunet.io/tos
 
@@ -41,8 +41,6 @@ func main() {
 
 	// Start listening for new messages coming via adapter
 	go adapter.StartMessageReceiver()
-	// Process deployment request worker
-	go messaging.DeploymentWorker()
 
 	// wait for server to start properly before sending requests below
 	time.Sleep(time.Second * 5)
@@ -59,6 +57,9 @@ func main() {
 
 	// get managed VMs, assume previous run left some VM running
 	firecracker.RunPreviouslyRunningVMs()
+
+	// Recreate host with previous keys
+	libp2p.CheckOnboarding()
 	wg.Wait()
 }