+++ models/machine.go
@@ -63,3 +63,9 @@ type ServiceResourceRequirements struct {
 	VCPU int
 	HDD  int
 }
+
+type Libp2pInfo struct {
+	ID         uint   `json:"id"`
+	PrivateKey []byte `json:"private_key"`
+	PublicKey  []byte `json:"public_key"`
+}